Fail differently when Parse.com credentials are missing.
The build failure introduced in 91e259b338809fe2bee10bae37939b879535471b was
tidy, but as it turns out, having the Gradle build fail outright renders
Android Studio unusable, and Android Studio doesn't seem to have a convenient
way to set the build environment. This commit changes things so that the app
gets assembled all right, but crashes hard at startup if the credentials are
not there.