Improve Android Instrumentation suite.
Use the Instrumentation.checkMonitorHit() method that I previously didn't know
about to verify that the external Activities have been started. This eliminates
the need for @FlakyTest annotations on those test methods.

Use viewAssertions instead of getActionBar() to determine the subtitle of the
ActionBar. This is more in the spirit of the testing framework, as you're not
really supposed to interact with the Activity instance. Unfortunately, the
@FlakyTest annotations cannot be removed; the test still runs unreliably
without them.