Set a white background in the application theme.
This causes Android to generate a startup screen that's the same color as our
app. Previously, it generated a black screen which we would then violently turn
white.

This background causes overdraw, because we must have a background on the
sliding fragment for the animation to work right. The overdraw problem is
solved by setting a transparent background on the application window in
onCreate() once we've started.