Opens details screens in separate activity.
Cleans up the settings fragment to no longer use a nested fragment. Gets rid of
the accessibility announcements because an Activity boundary is announced
anyway. Removes the unused animation resources that became obsolete in
d07017e82705e07e6ae4ff73197956781aedc061.

MainActivity can be further cleaned up by getting rid of the Handler/Message
stuff for the notification response. We can now just set a `PendingIntent` in
the notification that points to the detail activity with the correct MODE
parameter and back stack.

Toolbar styling needs more work, but is passable for now.